Refactor shared API client boundary
