+++ src/cli/remote/control-plane-client.ts
@@ -1,13 +1,8 @@
-import { createTRPCProxyClient, httpBatchLink } from '@trpc/client';
-import type { AppRouter } from '../../server/router.js';
 import type { ResolvedRuntimeHost } from '@/core/runtime/daemon/index.js';
+import { createControlPlaneProxyClient } from '@/client-shared/api/proxy.js';
 
 export function createDaemonControlPlaneClient(host: Extract<ResolvedRuntimeHost, { kind: 'daemon' }>) {
-  return createTRPCProxyClient<AppRouter>({
-    links: [
-      httpBatchLink({
-        url: `http://${host.endpoint.host}:${host.endpoint.port}/trpc`,
-      }),
-    ],
+  return createControlPlaneProxyClient({
+    url: `http://${host.endpoint.host}:${host.endpoint.port}/trpc`,
   });
 }
