+++ src/web-v2/api/client.ts
@@ -1,57 +1,5 @@
-import {
-  httpLink,
-  httpSubscriptionLink,
-  splitLink,
-} from '@trpc/client';
-import { createTRPCReact } from '@trpc/react-query';
-import type { inferRouterInputs, inferRouterOutputs } from '@trpc/server';
-import type { AppRouter } from '@/server/router';
-
-const trpcLinks = [
-  splitLink({
-    condition: (operation) => operation.type === 'subscription',
-    true: httpSubscriptionLink({
-      url: '/trpc',
-    }),
-    false: httpLink({
-      url: '/trpc',
-    }),
-  }),
-];
-
-export const trpcReact = createTRPCReact<AppRouter>();
-
-export function createControlPlaneTrpcClient() {
-  return trpcReact.createClient({
-    links: trpcLinks,
-  });
-}
-
-export type RouterInputs = inferRouterInputs<AppRouter>;
-export type RouterOutputs = inferRouterOutputs<AppRouter>;
-type AsyncIterableValue<T> = T extends AsyncIterable<infer Value> ? Value : T;
-export type ControlPlaneState = RouterOutputs['controlPlane']['state'];
-export type ControlPlaneSessionDetail = RouterOutputs['controlPlane']['session'];
-export type ControlPlaneSessionEventEnvelope = AsyncIterableValue<RouterOutputs['controlPlane']['sessionEvents']>;
-export type ControlPlaneSessionsEventEnvelope = AsyncIterableValue<RouterOutputs['controlPlane']['sessionsEvents']>;
-export type ControlPlaneHeartbeatEventEnvelope = AsyncIterableValue<RouterOutputs['controlPlane']['heartbeatEvents']>;
-export type ControlPlaneSessionMessage = NonNullable<ControlPlaneSessionDetail>['messages'][number];
-export type ControlPlanePendingApproval = RouterOutputs['controlPlane']['sessionPendingApproval'];
-export type ControlPlaneApprovalDecision = RouterInputs['controlPlane']['sessionResolveApproval']['decision'];
-export type ControlPlaneSessionSendPromptResult = RouterOutputs['controlPlane']['sessionSendPrompt'];
-export type ControlPlaneModelOptions = RouterOutputs['controlPlane']['modelOptions'];
-export type ControlPlaneSessionSettingsInput = RouterInputs['controlPlane']['sessionSettingsUpdate'];
-export type ControlPlaneHeartbeatTasks = RouterOutputs['controlPlane']['heartbeatTasks'];
-export type ControlPlaneHeartbeatTask = RouterOutputs['controlPlane']['heartbeatTask'];
-export type ControlPlaneHeartbeatTaskCreate = RouterOutputs['controlPlane']['heartbeatTaskCreate'];
-export type ControlPlaneHeartbeatTaskUpdate = RouterOutputs['controlPlane']['heartbeatTaskUpdate'];
-export type ControlPlaneHeartbeatTaskDelete = RouterOutputs['controlPlane']['heartbeatTaskDelete'];
-export type ControlPlaneHeartbeatTaskResume = RouterOutputs['controlPlane']['heartbeatTaskResume'];
-export type ControlPlaneHeartbeatTaskView = ControlPlaneHeartbeatTasks['tasks'][number];
-export type ControlPlaneHeartbeatRunView = ControlPlaneHeartbeatTask['runs'][number];
-export type ControlPlaneHeartbeatRun = RouterOutputs['controlPlane']['heartbeatRun'];
-export type ControlPlaneHeartbeatTaskRunNow = RouterOutputs['controlPlane']['heartbeatTaskRunNow'];
-export type ControlPlaneMemoryStatus = RouterOutputs['controlPlane']['memoryStatus'];
-export type ControlPlaneWorkspaceChanges = RouterOutputs['controlPlane']['workspaceChanges'];
-export type ControlPlaneWorkspaceChangedFile = ControlPlaneWorkspaceChanges['files'][number];
-export type ControlPlaneWorkspaceFileDiff = RouterOutputs['controlPlane']['workspaceFileDiff'];
+export {
+  createControlPlaneTrpcClient,
+  trpcReact,
+} from '@/client-shared/api/react';
+export type * from '@/client-shared/api/types';
